Code structure update
Created Icon manager that handles the click on icons to replace the original icon button listeners

Added a link of its prefab to tower_script

Added getters and setters and initiator of the ‘selectedStatus’ and relating fields of Player class

Rewrite the BuildTower function. It is now a member function of MapController and accepts two parameters, and is called in ‘onMouseDown’ function call of each grid. Condition checks are done in grid and iconManager classes.

TowerBuild script and TowerBuilder is now deprecated.

Minor code style adjustments

TODO:Fill other conditions of mouse clicks and update the selectedStatus.